+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # changelog
 
 ## [unreleased]
+* added Display and Error impls for Error
 * added examples to docs
 * improved test consistency